Move Read/WriteJSONFile to stringutil
Moving these funcs out of buildmodel represents their broad purpose better. But more importantly, it lets us reuse them without creating a dependency cycle between the buildmodel and sync packages.
