+++ stringutil/stringutil.go
@@ -3,7 +3,12 @@
 
 package stringutil
 
-import "strings"
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+	"strings"
+)
 
 // CutPrefix behaves like strings.Cut, but only cuts a prefix, not anywhere in the string.
 func CutPrefix(s, prefix string) (after string, found bool) {
@@ -11,4 +16,43 @@ func CutPrefix(s, prefix string) (after string, found bool) {
 		return s[len(prefix):], true
 	}
 	return s, false
+}
+
+// ReadJSONFile reads one JSON value from the specified file.
+func ReadJSONFile(path string, i interface{}) (err error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return fmt.Errorf("unable to open JSON file %v for reading: %w", path, err)
+	}
+	defer func() {
+		if closeErr := f.Close(); err == nil {
+			err = closeErr
+		}
+	}()
+
+	d := json.NewDecoder(f)
+	if err := d.Decode(i); err != nil {
+		return fmt.Errorf("unable to decode JSON file %v: %w", path, err)
+	}
+	return nil
+}
+
+// WriteJSONFile writes one specified value to a file as indented JSON with a trailing newline.
+func WriteJSONFile(path string, i interface{}) (err error) {
+	f, err := os.Create(path)
+	if err != nil {
+		return fmt.Errorf("unable to open JSON file %v for writing: %w", path, err)
+	}
+	defer func() {
+		if closeErr := f.Close(); err == nil {
+			err = closeErr
+		}
+	}()
+
+	d := json.NewEncoder(f)
+	d.SetIndent("", "  ")
+	if err := d.Encode(i); err != nil {
+		return fmt.Errorf("unable to encode model into JSON file %v: %w", path, err)
+	}
+	return nil
 }
